+++ signals/bb_width.py
@@ -17,7 +17,7 @@ from dateutil.relativedelta import relativedelta
 import time
 import utils.telegram as telegram
 import docs.database.database as database
-import utils.exchange as exchange
+import exchanges.binance as binance
 import ta
 
 def get_data(symbol, time_frame, start_date):
@@ -28,7 +28,7 @@ def get_data(symbol, time_frame, start_date):
 
     while retry_count < max_retry and not success:
         try:
-            df = pd.DataFrame(exchange.client.get_historical_klines(Symbol,
+            df = pd.DataFrame(binance.client.get_historical_klines(Symbol,
                                                                     time_frame,
                                                                     start_date
                                                                     ))
@@ -70,7 +70,7 @@ def apply_technicals(df, rsi_length):
 
 def super_rsi(symbol):
     # 15min timeframe
-    time_frame = exchange.client.KLINE_INTERVAL_15MINUTE
+    time_frame = binance.client.KLINE_INTERVAL_15MINUTE
     rsi_lookback_periods = 14 # 14 days
 
     # get start date