+++ signals/multi_emas.py
@@ -20,7 +20,7 @@ import time
 import utils.config as config
 import utils.telegram as telegram
 import utils.database as database 
-import utils.exchange as exchange
+import exchanges.binance as binance
 
 TELEGRAM_PREFIX_SIGNAL = "MULTI-EMAs" 
 
@@ -33,7 +33,7 @@ def get_data(symbol, time_frame, start_date):
 
     while retry_count < max_retry and not success:
         try:
-            df = pd.DataFrame(exchange.client.get_historical_klines(symbol,
+            df = pd.DataFrame(binance.client.get_historical_klines(symbol,
                                                                     time_frame,
                                                                     start_date
                                                                     ))
@@ -78,7 +78,7 @@ def apply_technicals(df):
 
 def main(symbol):
     # 1D timeframe
-    time_frame = exchange.client.KLINE_INTERVAL_1DAY
+    time_frame = binance.client.KLINE_INTERVAL_1DAY
 
     # get start date
     # get max data as possible to make sure the slowest emas 200 and 300 get the same values as tradingview 
@@ -203,7 +203,7 @@ def get_symbols(trade_against):
     # Get blacklist
     # blacklist = get_blacklist()
 
-    exchange_info = exchange.get_exchange_info()
+    exchange_info = binance.get_exchange_info()
 
     symbols = set()